I think I have cleared up some misunderstandings
The array factor is a complex scalar quantity. It is the factor that
multiplies into the element factor.

The element factor describes how the dipoles respond to x,y/theta,phi
components of an incoming electric field. The array factor describes the
gain and phase response of the whole station?? Take everything here with
a grain of salt